[UPD] Required properties with non-reference data type in the DTOs classes were made as null in order to apply correctly the RequiredField annotation.            First, this allow null values for the data type, then the RequiredField annotation validates that a not null value has been sent.

+++ Domain/DTOs/Response/CommentResponse.cs
@@ -8,7 +8,7 @@ namespace Domain.DTOs.Response
         public Guid CourseId { get; set; }
         public string StudentName { get; set; }
         public string Message { get; set; }
-        public int Score { get; set; }
+        public byte Score { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
         public Guid? CreatedBy { get; set; }
